moved webtests to separate files and directory
