+++ tests/test_query.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from contextlib import nullcontext as does_not_raise
 import datetime
-import os
 
 import pytest
 
@@ -9,30 +8,6 @@ from requests import Request
 
 from bookops_worldcat.errors import WorldcatRequestError
 from bookops_worldcat.query import Query
-from bookops_worldcat.metadata_api import MetadataSession, WorldcatAccessToken
-
-
-@pytest.mark.webtest
-def test_query_live(live_keys):
-    token = WorldcatAccessToken(
-        key=os.getenv("WCKey"),
-        secret=os.getenv("WCSecret"),
-        scopes=os.getenv("WCScopes"),
-    )
-    with MetadataSession(authorization=token) as session:
-        header = {"Accept": "application/json"}
-        url = "https://metadata.api.oclc.org/worldcat/search/brief-bibs/41266045"
-        req = Request(
-            "GET",
-            url,
-            headers=header,
-        )
-        prepped = session.prepare_request(req)
-
-        with does_not_raise():
-            query = Query(session, prepped, timeout=5)
-
-        assert query.response.status_code == 200
 
 
 def test_query_not_prepared_request(stub_session):
